core: centralize the derivation of GB conflict metadata
Deduplicate the code related to the extraction of GitButler conflict
metadata out of gix merge outcomes.
Share it across rebase, and gitbutler-repo.
